requirements.txt
# Файл залежностей для торгового бота
ccxt>=4.0.0
aiohttp>=3.8.0
numpy>=1.24.0
pandas>=2.0.0
websockets>=11.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
asyncio-mqtt>=0.16.0
requests>=2.31.0
python-telegram-bot>=20.0.0
schedule>=1.2.0
ta>=0.10.2
plotly>=5.17.0
streamlit>=1.28.0
